refactor: remove duplicate import statement to hooks
