+++ src/domains/help/pages/Article/Article.tsx
@@ -1,7 +1,7 @@
 import { Button } from "app/components/Button";
 import { Icon } from "app/components/Icon";
 import { Page, Section } from "app/components/Layout";
-import { useActiveProfile } from "app/hooks/env";
+import { useActiveProfile } from "app/hooks";
 import React from "react";
 import { useTranslation } from "react-i18next";
 import { styled } from "twin.macro";
