+++ src/domains/plugin/pages/PluginsCategory/PluginsCategory.tsx
@@ -5,7 +5,7 @@ import { Icon } from "app/components/Icon";
 import { Page, Section } from "app/components/Layout";
 import { LayoutControls } from "app/components/LayoutControls";
 import { SearchBarPluginFilters } from "app/components/SearchBar/SearchBarPluginFilters";
-import { useActiveProfile } from "app/hooks/env";
+import { useActiveProfile } from "app/hooks";
 import { InstallPlugin } from "domains/plugin/components/InstallPlugin";
 import { PluginGrid } from "domains/plugin/components/PluginGrid";
 import { PluginList } from "domains/plugin/components/PluginList";
