+++ src/domains/transaction/pages/SendTransfer/SendTransfer.tsx
@@ -6,9 +6,7 @@ import { Page, Section } from "app/components/Layout";
 import { StepIndicator } from "app/components/StepIndicator";
 import { TabPanel, Tabs } from "app/components/Tabs";
 import { useEnvironmentContext } from "app/contexts";
-import { useClipboard } from "app/hooks";
-import { useActiveProfile, useActiveWallet } from "app/hooks/env";
-import { useValidation } from "app/hooks/validations";
+import { useActiveProfile, useActiveWallet, useClipboard, useValidation } from "app/hooks";
 import { AuthenticationStep } from "domains/transaction/components/AuthenticationStep";
 import React, { useEffect, useMemo, useState } from "react";
 import { useForm } from "react-hook-form";
